refactor: OpenAIのAPIを叩く関数をutils以下に作成
